# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/plaid/me/link-token": {
      "post": {
        "description": "Returns a short-lived link_token. Pass it to the Plaid Link component on the frontend.",
        "operationId": "PlaidMeController_createLinkToken",
        "parameters": [],
        "responses": {
          "201": {
            "description": "link_token string"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a Plaid link_token to initialise Plaid Link",
        "tags": ["plaid"]
      }
    },
    "/plaid/me/exchange-token": {
      "post": {
        "description": "Call this after Plaid Link calls onSuccess with a public_token. The backend exchanges it for an access_token, persists the Item, and returns the Item record.",
        "operationId": "PlaidMeController_exchangeToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created Plaid Item (without access_token)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Exchange a public_token for a permanent access_token (Plaid Item)",
        "tags": ["plaid"]
      }
    },
    "/plaid/me/items": {
      "get": {
        "operationId": "PlaidMeController_getMyItems",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of Plaid Items (access_token omitted)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the authenticated user's linked Plaid Items",
        "tags": ["plaid"]
      }
    },
    "/plaid/me/accounts": {
      "get": {
        "operationId": "PlaidMeController_getMyAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of Plaid AccountBase objects"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the authenticated user's accounts across all linked Items",
        "tags": ["plaid"]
      }
    },
    "/plaid/me/transactions": {
      "get": {
        "description": "Returns transactions stored locally (synced via webhooks + hourly fallback). Also enqueues a sync job per linked Item so the next read is fresh. Use ?limit and ?before for pagination.",
        "operationId": "PlaidMeController_getMyTransactions",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 100,
              "type": "string"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "description": "ISO date — return transactions with date earlier than this",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of stored transactions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read the authenticated user's persisted transactions and trigger a background sync",
        "tags": ["plaid"]
      }
    },
    "/plaid/me/balance/refresh": {
      "post": {
        "description": "Calls /accounts/balance/get for each Item, which polls the institution for the latest available/current balances. More expensive than GET /plaid/me/accounts (which uses cached data). Use sparingly.",
        "operationId": "PlaidMeController_refreshBalances",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of accounts with refreshed balances"
          },
          "403": {
            "description": "No Plaid account linked"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Force a real-time balance refresh across all linked Items",
        "tags": ["plaid"]
      }
    },
    "/plaid/me/auth": {
      "get": {
        "description": "Calls /auth/get for each Item. Requires the auth product to be enabled on the Item. In sandbox the default test institution returns deterministic numbers.",
        "operationId": "PlaidMeController_getMyAuth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of per-item auth payloads (accounts + numbers)"
          },
          "403": {
            "description": "No Plaid account linked"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get routing and account numbers for the user's linked accounts",
        "tags": ["plaid"]
      }
    },
    "/plaid/me/identity": {
      "get": {
        "description": "Calls /identity/get for each Item. Requires the identity product to be enabled on the Item.",
        "operationId": "PlaidMeController_getMyIdentity",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of per-item identity payloads"
          },
          "403": {
            "description": "No Plaid account linked"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get account holder names, emails, phones, and addresses",
        "tags": ["plaid"]
      }
    },
    "/plaid/me/liabilities": {
      "get": {
        "description": "Calls /liabilities/get for each Item. Returns null per-item if the Item does not support liabilities.",
        "operationId": "PlaidMeController_getMyLiabilities",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of per-item liabilities payloads"
          },
          "403": {
            "description": "No Plaid account linked"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get liability details (credit cards, student loans, mortgages)",
        "tags": ["plaid"]
      }
    },
    "/plaid/me/investments/holdings": {
      "get": {
        "description": "Calls /investments/holdings/get for each Item. Returns null per-item if the Item does not support investments.",
        "operationId": "PlaidMeController_getMyHoldings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of per-item holdings payloads"
          },
          "403": {
            "description": "No Plaid account linked"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get investment account holdings (positions and securities)",
        "tags": ["plaid"]
      }
    },
    "/plaid/me/investments/transactions": {
      "get": {
        "description": "Calls /investments/transactions/get for each Item. Both startDate and endDate must be ISO YYYY-MM-DD. Defaults to the last 90 days.",
        "operationId": "PlaidMeController_getMyInvestmentTransactions",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "example": "2024-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "example": "2024-04-01",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of per-item investment transactions payloads"
          },
          "403": {
            "description": "No Plaid account linked"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get investment transactions for a date range",
        "tags": ["plaid"]
      }
    },
    "/plaid/me/items/{id}": {
      "delete": {
        "description": "Calls /item/remove on Plaid then deletes the Item from Flamingo. The `id` parameter is the Flamingo UUID of the PlaidItem record.",
        "operationId": "PlaidMeController_removeItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Item removed successfully"
          },
          "403": {
            "description": "No Plaid account linked"
          },
          "404": {
            "description": "Item not found or does not belong to this user"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Unlink a Plaid Item (disconnect a bank)",
        "tags": ["plaid"]
      }
    },
...
    "/plaid/items": {
      "get": {
        "description": "Returns every PlaidItem record; access_token is omitted.",
        "operationId": "PlaidAdminController_getAllItems",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of PlaidItem records (access_token omitted)"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "admin-jwt": []
          }
        ],
        "summary": "List all Plaid Items across all users (admin only)",
        "tags": ["plaid"]
      }
    },
    "/plaid/items/{id}": {
      "delete": {
        "description": "Calls /item/remove on Plaid (best-effort) and deletes the record from Flamingo. The `id` parameter is the Flamingo UUID of the PlaidItem record.",
        "operationId": "PlaidAdminController_removeItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Flamingo PlaidItem UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Item removed successfully"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "PlaidItem not found"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "admin-jwt": []
          }
        ],
        "summary": "Force-remove a Plaid Item (admin only)",
        "tags": ["plaid"]
      }
    },
...
  },
  "info": {
    "title": "Flamingo API",
    "description": "Flamingo REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "admin-jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ExchangeTokenDto": {
        "type": "object",
        "properties": {
          "publicToken": {
            "type": "string",
            "description": "The public_token returned by Plaid Link onSuccess callback",
            "example": "public-sandbox-xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
          }
        },
        "required": ["publicToken"]
      },
...
    }
  }
}